I updated the LICENSE and README.md.
In the LICENSE, I included the notice that I put at the top of
each of the Corundum code files (slightly modified). It may be a good
idea to change all the comments in the code to reflect that small change.

In the README.md, I included a short description of Corundum and a very
short set of generalized installation instructions.

Signed-off-by: REALDrummer <conman.davey@gmail.com>