I deleted myWiki and added lots to MessageColor.
I deleted myWiki because it will not be useful any longer. Its data will be
distributed among the final variables and enumerated types and such through
the Corundum A.P.I.

I made MessageColor its own class because it was growing large and becuase
it should be more accessible. I also added Javadoc documentation to every-
thing in MessageColor.java and created some new methods that will be
essential in the use of the MessageColor class.

Signed-off-by: REALDrummer <conman.davey@gmail.com>