I added Zone and BoundingBox outlines and finished BlockType.
Firstly, I'm declaring BlockType complete and fully documented and ready
to go. There may be more utility methods or whatnot that we want to add
later, and those can be added accordingly, but for now, this is done.

Next, I made Zone and BoundingBox classes to describe zones in the
Minecraft world and the bounding boxes of blocks and entities, respectively.
Both are only outlines for now. I used the BoundingBox in a utility
method in BlockType.

Finally, I removed some stuff from BiomeType temporarily just to make the
rest of the code compilable.

Signed-off-by: REALDrummer <conman.davey@gmail.com>