I moved things around, added secure(), handled plugins, improved Messenger, and integrated Logger.
Firstly, I moved lots of things around in packages. I got rid of the
Minecraft package level because it was rather pointless. I also rearranged
some classes in order to restrict visibility from the A.P.I.

Also related to A.P.I. visibility, I created this Corundum.secure() method
that uses stack traces to basically attempt to ensure that the plugin
authors can't use certain methods that might have to be public. However,
I advise just rearranging things in packages if possible to avoid having
to use this method, which is slower and might be ignored by authors who
don't read the Javadoc.

I also implemented the plugin handling methods like load() and disable(),
though they need to be tested. Also, enable(), disable(), and unload()
still need to generate plugin-handling listener events and be cancellable.

Messenger is much more complete now, but still needs some work before it's
finished. I made a utils subpackage specific to messaging to aid in
ensuring that certain methods are only accessible where they're needed.

Finally, I figured out that for console/logging messages, Minecraft just
uses Apache's Log4J, so I implemented messaging for Console with that.
I know already that this will cause problems with console colors, which
have an escape sequence in the console, but will appear in the logs as
text. It's been a bug in Minecraft for a long time. I won't worry about
it for now, but some day, I hope we could fix that.

Signed-off-by: REALDrummer <conman.davey@gmail.com>