I added to the Recipes.
I began working on the Enchantment class, when I realized I had to define
Enchantments with the ItemTypes they work on, so I started working on
ItemTypes, which is when I realized I needed to give the ItemTypes recipes,
so I started working on Recipes.

I made CraftingRecipe an abstract class so that we can have "shaped" and
"shapeless" recipes handled with separate classes with different
constructors for each.

I also made a constructor for BrewingRecipe and more constructors for Crafting-
Recipe. I will be adding more constructors for CraftingRecipe not just for
the A.P.I., but also for use in ItemType. As it turns out, I can't define
ItemTypes with ItemTypes that weren't defined earlier, so I'll have to
make Recipe constructors that can take in item I.D.s and construct the
types from there.

Signed-off-by: REALDrummer <conman.davey@gmail.com>