I added more stuff to BiomeType and got rid of I.D.s in the constructors.
I looked through the biome I.D. list and noticed that the first 40 are
all the regular biomes with no skipping I.D.s and that all the variant
biomes had their parents' I.D. + 128.

I also added constructor parameters to all the enumerated types and got
rid of the temporary BiomeType default constructor.

I also made a couple small modifications to ItemType's and BlockType's
getByID methods to 1) clarify in the Javadoc that they return null if no
matches are found and 2) to make them slightly more efficient by cancelling
the search completely if the I.D. is < 0.

Signed-off-by: REALDrummer <conman.davey@gmail.com>