I finished eliminating declared materials from BiomeType.
I figured out how to get all the properties we had in BiomeType already
using Minecraft's code (though we need to figure out how to process the
spawning animal types). I also made all the enum constructors much shorter
and ordered them.

The BiomeOE class ("OE" for "overextension") is there to allow us to access
protected fields and methods in Minecraft's biome package; however, I'm
not sure if there are legal or coding issues with having the same package
as Minecraft's server jar, so we should replace it with true overextension
(as I described in the BiomeOE class's Javadoc).

Signed-off-by: REALDrummer <conman.davey@gmail.com>