I added Holdable and finished enumerating BiomeTypes.
I added the last few BiomeTypes to the enum list. That was pretty simple.

I added the Holdable interface, too. It's to be used for everything that
can be held in an inventory, namely Items and Blocks. I took almost all
the methods from IDedType and put them in Holdable so that things that are
I.D.ed like BiomeTypes can still implement this interface, but not have
the unnecessary methods like data value stuff (since biomes don't have
data values, only I.D.s).

Signed-off-by: REALDrummer <conman.davey@gmail.com>