I added ALL the ItemType enumerations.
Yes, I added them all. They're all there now, including all the potions and
all the new 1.8 items. Everything is there.

On a more minor note, I also had to change the getData() method in ItemType
(plus the interface HoldableType and BlockType) to return a short instead
of a byte because potion data values are too big.

Also, I removed that list of changes between the site and the enumerations
from the Javadoc in BlockType and didn't put any in ItemType. I actually
used a different site for some of the ItemTypes because the site I usually
use removed all the potions, plus I didn't think anyone would really care
that much and it just cluttered the Javadoc.

MAN, we got a lot done today! I love weekends!

Signed-off-by: REALDrummer <conman.davey@gmail.com>