I started the SettingsManager class and made Configurable.
As discussed in the comments on commit b268fef, this SettingsManager can
be used to manage server settings and player settings and more and handle
the automatic loading and saving of said settings using the JSON parser
built into Minecraft already.

Configurable is an interface that basically tells us that this Object
has a SettingsManager. Players, CorundumServers, and some sort of
player group Object should implement this interface.

Signed-off-by: REALDrummer <conman.davey@gmail.com>