I added some more things to SettingsManager.
I added a File instance variable to represent the file that this Settings-
Manager can load from and save to.

Also, I added a "parent" SettingsManager that can allow the Settings-
Manager to look in its parents' settings if it does not contain a specific
setting. We can use this to allow players' settings to inherit from group
settings which inherit from server settings, allowing us to make a system
where any given property can be found in the server settings, but settings
can be changed for different groups or players on a more individual basis.
However, I have not yet implemented this parent setup fully.

Signed-off-by: REALDrummer <conman.davey@gmail.com>