I added some stuff to SettingsManager and fixed init problems.
I added some constructors and utils to SettingsManager and implemented
SettingsManager inheritance. I didn't see that Niadel had already done it
until afterward....

Also, the argInfo setting stuff in the startServer() method doesn't work
because it calls fields in MinecraftServer that are not yet initialized;
see the comments I made for more details.

Signed-off-by: REALDrummer <conman.davey@gmail.com>