I fixed some merge issues and removed the parent settings retrieval.
The local merge added in both new SettingsManager(File, Object...) con-
structors and load() and save() methods. Both constructors were the same,
so I just took out one; I took out my versions of the load() and save()
methods because they were simply skeletons while the remote versions were
actually implemented.

Also, I removed the part of the code in the constructor that retrieves the
settings from the parent because instead, the SettingsManager retrieves
parent settings when they are called or set. This may be (very slightly)
less efficient performance-wise, but it saves huge amount of memory by
avoiding having to have copies of every single setting available in all
the plugins for every single player and group of players when most of the
time, players or groups will simply have the same settings as their parent
(groups or the server, respectively).

Signed-off-by: REALDrummer <conman.davey@gmail.com>