I re-added the code to download the Minecraft server jar in the launcher.
I was messing with the launcher trying to solve the issue concerning the
GUI logging and accidentally made the code Niadel wrote that downloads the
Minecraft_server.jar disappear. I put it back the way it was.

Also, I had to get rid of the final CORUNDUM_GUI_ENABLED field. It's not
meant to be final and it was creating areas of dead code because of that.
I understand that it was put in that way because the Corundum GUI can't be
enabled until we build it, but we should instead just initialize it as
false and not change it for now; making it final causes dead code in areas
that we will eventually use when the Corundum GUI is made.

Signed-off-by: REALDrummer <conman.davey@gmail.com>