I am attempting to rework the type system for inheritance.
I realized looking through the Minecraft code that lots of Minecraft type
parts -- especially Entities -- will have to be divided into groups in
many different ways. If we kept using enums for all our types, then we
would have to remake all the methods from each inherited type and come up
with some kind of parent-child relationships and all sorts of crazy stuff
for things like MobTypes, which are all derived from EntityTypes. So,
instead, I decided to embark on a journey to rework the typing system using
a kind of pseudo-enum system (with values() and all that implemented in
abstract classes and enum values implemented as public static final
variables). That way, we will be able to make our types inherit properties
from more general types so MobTypes can inherit properties from EntityTypes,
for example.

There are still bugs here and this build is not yet compilable. That's why
I put it in a separate branch. If it fails, we still have the rest of the
old code. If it works, I'll merge it into master when it's done.

Signed-off-by: REALDrummer <conman.davey@gmail.com>