[WIP] I've continued work on making Type classes extendable.
I was having trouble figuring out a good way to store and handle value sets
for each subclass since the values would have to be static, but for each
subclass, which is problematic since static things can't be inherited.
I resolved it by creating a static HashMap<Class, IDedType[]> in
IDedType.class. Now each subclass can have its own array of appropriate
values associated with its class. Java Reflection can be so useful.

As the "[WIP]" tag indicates, this is still a work in progress and it
should be noted that there are still compilation errors with this setup.
I'm working to fix it.

Signed-off-by: REALDrummer <conman.davey@gmail.com>