[WIP] I finished IDedType and IDedTypeWithData and started converting types.
I worked out all the issues with IDedType and IDedTypeWithData abstract
classes. I'm still in the process of cconverting all the Type classes to
the new format, but it shouldn't be that difficult, especially thanks to
Eclipse's regex search and replace that lets me use only two "Replace All"s
to convert all the enumerated types to public static final constructor
calls.

Signed-off-by: REALDrummer <conman.davey@gmail.com>