I added constructors for PluginThread and moved it to Corundum.plugins.
I added constructors that emulate all the regular Thread constructors,
which should hopefully mean that we can redirect all Thread constructions
to PluginThread constructors for any given Thread constructor called in
a plugin handled by a PluginClassLoader.

I implemented a method to get the plugin from the current thread if it's a
PluginThread, but of course, this still needs to be tested.

Also, I moved it to the "plugins" package with the CorundumPlugin class so
that we can let CorundumPlugin and PluginThread utilize default-visible
and protected constructors and fields and whatnot.

Signed-off-by: REALDrummer <conman.davey@gmail.com>