[WIP] I'm attempting to fix GUI logging issue (issue #12).
I've been creating a pseudo-inheritance system to allow ClassLoaders to
effectively search multiple jars for a resource. ClassLoaders already have
parents, but I added children for downward recursion and implemented my
own parents in an attempt to prevent Log4J from getting access to the
highest level (system) ClassLoader. However, I've since realized that
the child might not be necessary except for the Minecraft server jar since
Log4J tries to use system ClassLoaders to load resources and that I may
not need to implement my own parents to "hide" them from Log4J if we take
over the system ClassLoader, which I discovered can be done with the Java
VM global variable `-Djava.system.class.loader=[your loader]`.

I'll work on making this more efficient later; for now, I'm just trying to
get it to work for now. I did realize that recursing down is actually
unnecessary as long as we make a new ClassLoader for each jar we load, so
I'll get rid of that soon enough (except for the minecraft_server.jar since
Log4J insists on trying to use parent and system ClassLoaders instead of
the one it's given.

Signed-off-by: REALDrummer <conman.davey@gmail.com>