I made lots of rather minor changes.
I made TypedObject an interface because everything was abstract anyway.

I made a new interface called PhysicalObject that lies between TypedObject
and Holdable to implement Enchantments better; it no longer tries to
require that Enchantments have a Location.

I added more methods to AbstractCorundumServer and renamed it "Server"
because it's shorter and the "Abstract" made it sound like an Abstract
class when it's really an interface.

I added a few more things to Item, including a couple more constructors.

I got rid of a lot of the Minecraft object retrieval methods, e.g.
`getMCEnchant()` in EnchantmentType. In order to make the A.P.I. look nice
and for security reasons, we have to ensure that these things are not
visible. So far, I have been unable to find a way to ensure that they are
invisible to the plugin authors without making them invisible to pretty
much all of Corundum, including the package that they're in, so I just
got rid of them. If we can ensure A.P.I. invisibility and usefulness, we
can add them back, but we need to ensure those two things.

Signed-off-by: REALDrummer <conman.davey@gmail.com>