[WIP] I'm working to reengineer the EntityTypes to make them work better.
As they were, EntityTypes had some issues. The main reason for this
revision is to set it up so that parent types simply point to their lowest
types, which means that if someone wants to check whether or not an
EntityType is a MonsterType, all they have to do is use `instanceof`!
What's more, with this, I solved an issue with not being able to properly
add EntityType parent types to the values map in IDedType without writing
a whole lot of ugly code.

Also, on a more minor note, with this system, plugin authors can make new
types of Entities simply using `new Entity([any EntityType])`!

Signed-off-by: REALDrummer <conman.davey@gmail.com>