[WIP] I added more NonLivingEntity classes and made Player a LivingEntity.
I'm continuing to make more Entity classes to work with so that the A.P.I.
will work properly. Sadly, the way this is set up, the compilation errors
will not end until we have all of these classes done.

Also, I uncommented the "Player extends LivingEntity" statement. I took out
the Player class's `playerMC` since the Entity class's `entityMC` should
be used to store the Minecraft version of the object instead. However, this
caused errors since `playerMC` was an `EntityPlayerMC` and `entityMC` is
just a plain old Minecraft `Entity`. To avoid having to cast all this
stuff, what I think I'll do is self-parameterize the Entity classes and
use the parameterized type to make a `getEntityMC()` method that return
`entityMC` as the appropriate type.

Signed-off-by: REALDrummer <conman.davey@gmail.com>