[WIP] I added more EntityTypes and added some stuff to myList.
For one thing, I added a PlayerEntitiy class that's meant to represent the
Player's Entity. Normally this would just be used to represent Players
in-game, but it's meant to also be usable for creating new fake Players
like plugins that can create guards outside of towns.

I also added more EntityTypes and their appropriate classes. I am now done
with the NonLivingEntities; I'm working on the LivingEntities now. Boy,
this is going to take a lot of work for Blocks and Items....

I've been posting about Corundum on Minecraft-related sub-Reddits and the
like. We really need more developers; this is not going as quickly as I
would have hoped. I blame myself a lot for that; I decided I wanted to
start a town on my server at the beginning of break and now that's cutting
a lot into time I could have used to work on Corundum. Still, I'm getting
stuff done and I hope we can get more developers on board with this
project.

Signed-off-by: REALDrummer <conman.davey@gmail.com>