[WIP] I implemented all the Animal subclasses.
There's one issue with the Horse class caused by me not realizing that
EntityHorse does not extend EntityTameable, which seems stupid to me since
horses are definitely tameable... though I guess not in the same way as
ocelots and wolves. Anyway, I'll fix that.

Now all that's left in the entities is the Monsters.

Signed-off-by: REALDrummer <conman.davey@gmail.com>