[WIP] I added an `EntityType.fromMC()` method and `FlyingMonster`s.
I was making the `Monster` class `implement` `HostileMob` when I realized
that we had no system of direct generalized translation from a Minecraft
object to a Corundum object. We have constructors to make the Minecraft
Objects into Corundum Objects, but nothing that would allow us to take
a Minecraft Object and directly translate it into a Corundum Object. This
`fromMC()` method works just like `create()`, but rather than call the
nullary constructor for a Corundum Entity, it calls the unary constrcutor
with an MC argument. I already implemented it for all the base classes
created so far.

I also added the `FlyingMonster` class and its bases, `Blaze` and `Ghast`.

Signed-off-by: REALDrummer <conman.davey@gmail.com>