The Entity classes are all done!
I finished adding the remainder of the Monster classes. The only missing
pieces are the I.D.s for the uncategorized `Monster`s (`Creeper`,
`Enderman`, and `Witch`) and there may have been a couple monsters I
missed; as I write this commit, the WiFi is down. I will make any necessary
corrections when the WiFi is usable again and I can check Minepedia and
push this commit.

I also updated Biome to fix compilation errors, so now no compilations
errors exist in the project. Prepare for a merge of this branch!

Signed-off-by: REALDrummer <conman.davey@gmail.com>