I added `getName()`, made `CreatableType` an interface, and fixed `Entity`.
I added the `getName()` method to be able to get the name of any `IDedType`
and implemented it in the `Block`, `Item`, and `Entity` classes (plus a
few other inner classes that extended `IDedType` like `Profession` in
`Villager`).

I also made CreatableType an interface since it didn't add anything on its
own and it causes less inheritance conflict, especially with a new
interface I created for the things that have Minecraft object equivalents,
`MCEquivalentTyped`.

Finally, I fixed some problems in `Entity`, mostly pertaining to the
`Location`, `Velocity`, and `Rotation` objects that we cannot use if we
want to avoid a lot of confusion.

Signed-off-by: REALDrummer <conman.davey@gmail.com>