I reorganized some utility classes/interfaces and improved `World`.
I moved that `types` package into the `utils` package and moved some of
the former classes that were converted to interfaces into the
`utils.interfaces` package.

Also, I removed some constructors from `World` that would have caused
a million bugs and. I also made `World` (and `Entity`) implement the new
interface `MCEquivalent` and changed the names of a couple methods that
were already there to fit with this interface.

Signed-off-by: REALDrummer <conman.davey@gmail.com>