I made some minor improvements.
Firstly, I removed the EnchantmentExceptions from the Enchantment class.
Plugin developers should be able to create illegal enchantments if they
want; if they want to make sure the enchantments are legal, we can create
utilities for that.

I also added a configurable name and message prefix to the CorundumServer
class. (I also added getters and setters for these to the Server interface
so that the Corundum Hub can access them.)

Finally, I made some other minor improvements here and there.

Signed-off-by: REALDrummer <conman.davey@gmail.com>