I added implementation of the Biome object.
Biome objects were tricky because they have no Minecraft equivalent and
bounding them would be difficult and C.P.U.-intensive. For now, I figured
what we could do is simply keep track of a single location inside the
biome; if someone wanted to know if, say, another location was in the same
biome, we could use this location alone to try to find a path between the
two points that stays in the same biome type. In short, I figure that one
location in the biome might be enough to do any sort of calculations we
may need and takes very little C.P.U. and memory most of the time.

Signed-off-by: REALDrummer <conman.davey@gmail.com>