I added utils to Biome and prepared Block for its reenginnering.
I began reworking Block to match what we did with Entity and its
subclasses.

In Biome, I added a utility to get a "circumscribing zone", a zone that is
as small as possible while still containing the whole contiguous biome. It
ended up being an exciting challenge for some SoDA members and I to come
up with efficient algorithms for this. I implemented an edge-following
algorithm for it, but there was also a graph search idea and a midpoint
idea that have yet to be explored. Maybe at some point we'll implement the
other two and test and compare their performance on some real Minecraft
maps.

I also created an ArmorMaterial interface that will be useful for Items
and Blocks that armor can be made of like leather or diamonds. I'll make
a similar interface for ToolMaterials like diamonds and cobblestone.

I also touched up here and there and made minor improvements.

Signed-off-by: REALDrummer <conman.davey@gmail.com>