I fixed the "Compile the source" subheader.
I noticed that the "Compile the source" subheader in README.md had not formatted properly and treated the "####" as a literal. I should have noticed sooner. I fixed that.

I also attempted to fix the way the numbering restarts after each subheader, but I don't think that's an easy fix. I'm not sure how to do it at the moment. In the plain text document, the numbers are correct, but GitHub's markdown formatting changes the numbers.