I created and documented Command and created Parser. [WIP]
I have most of the groundwork laid for Parser. I'm in the process of
adding more standard Parsers, but that requires some other things to be
done in the A.P.I. for much of that to work, e.g. getting current online
Players and their names.

Command has been fully documented and I have created some utility classes
for it, but Command itself is rather skeletal as of yet.

There are also some changes in progress in the blocks type system. I really
shouldn't mix modular changes like that; I'll get that dealt with soon.

Signed-off-by: REALDrummer <conman.davey@gmail.com>